shared/entity: Fix imports from reverting moving the entity package.
