+++ lxd/metrics/metrics_test.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/canonical/lxd/lxd/entity"
 	"github.com/canonical/lxd/shared/api"
+	"github.com/canonical/lxd/shared/entity"
 )
 
 func TestMetricSet_FilterSamples(t *testing.T) {
